refactor: move review modal outside scope
