+++ apps/web/components/ui/ReviewStatistics/ReviewStatistics.vue
@@ -36,37 +36,10 @@
       </div>
     </div>
   </div>
-
-  <UiModal
-    v-if="isReviewModalOpen"
-    v-model="isReviewModalOpen"
-    aria-labelledby="review-modal"
-    tag="section"
-    role="dialog"
-    class="h-full md:w-[500px] md:h-fit m-0 p-0"
-    :data-testid="dataTestId"
-  >
-    <header>
-      <h3 class="font-bold typography-headline-4">
-        {{ t('review.createReviewFormTitle') }}
-      </h3>
-      <UiButton @click="closeReviewModal" square variant="tertiary" class="absolute right-2 top-2">
-        <SfIconClose />
-      </UiButton>
-    </header>
-
-    <template v-if="isAuthorized">
-      <ReviewForm @on-close="closeReviewModal" class="h-fit" />
-    </template>
-    <template v-else>
-      <LoginComponent v-if="isLogin" @change-view="isLogin = false" />
-      <Register v-else @change-view="isLogin = true" />
-    </template>
-  </UiModal>
 </template>
 
 <script setup lang="ts">
-import { SfRating, SfIconClose, SfProgressLinear, SfIconStarFilled } from '@storefront-ui/vue';
+import { SfRating, SfProgressLinear, SfIconStarFilled } from '@storefront-ui/vue';
 import type { ReviewStatisticsProps } from './types';
 import { productGetters, reviewGetters } from '@plentymarkets/shop-api';
 
@@ -75,12 +48,8 @@ const props = defineProps<ReviewStatisticsProps>();
 const productId = Number(productGetters.getItemId(props.product));
 
 const { t } = useI18n();
-const { isAuthorized } = useCustomer();
 const { data: productReviewAverage } = useProductReviewAverage(productId);
-const { isReviewModalOpen, openReviewModal, closeReviewModal } = useProductReviews(productId);
-
-const isLogin = ref(true);
-const dataTestId = computed(() => (isAuthorized.value ? 'review' : 'login') + '-modal');
+const { openReviewModal } = useProductReviews(productId);
 
 const reviewAverageText = computed(() => reviewGetters.getAverageRating(productReviewAverage.value, 'tenth'));
 const reviewAverageStars = computed(() => reviewGetters.getAverageRating(productReviewAverage.value, 'half'));
